% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractClass.R
\name{AbstractClass}
\alias{AbstractClass}
\title{Create an abstract R6 Class}
\description{
Creates an abstract R6 class by placing a thin wrapper around \link[R6:R6Class]{R6::R6Class}
which causes an error to be thrown if the class is directly constructed
instead of one of its descendants.
}
\details{
All arguments of \link[R6:R6Class]{R6::R6Class} can be used as usual, see full
details at \link[R6:R6Class]{R6::R6Class}.
}
