% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R6.R
\name{is.R6}
\alias{is.R6}
\alias{assert_R6}
\title{Is 'x' a R6 object or class?}
\usage{
is.R6(x)

assert_R6(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Either TRUE/FALSE is testing if \code{x} inherits from
\code{R6} or \code{R6ClassGenerator}, otherwise returns \code{x} invisibly on assertion
if TRUE or returns an error if FALSE
}
\description{
Assert/test if 'x' is a R6 object or class
}
