% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R6.R
\name{is.R6Object}
\alias{is.R6Object}
\alias{assert_R6Object}
\title{Is 'x' a R6 object?}
\usage{
is.R6Object(x)

assert_R6Object(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Either TRUE/FALSE is testing if \code{x} inherits from \code{R6}, otherwise
returns \code{x} invisibly on assertion if TRUE or returns an error if FALSE
}
\description{
Assert/test if 'x' is a R6 object
}
