% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{private}
\alias{private}
\title{Get R6 object private environment}
\usage{
private(x)
}
\arguments{
\item{x}{(\code{R6}) \cr R6 object to get environment from, errors if not R6}
}
\description{
Access the private environment of an R6 object
}
