% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{super}
\alias{super}
\title{Get R6 object parent environment}
\usage{
super(x)
}
\arguments{
\item{x}{(\code{R6}) \cr R6 object to get environment from, errors if not R6}
}
\description{
Access the parent environment of an R6 object
}
